Rename test package to match production package for jobs report
